requirements.txt
fastapi
uvicorn
pillow 
faiss-cpu
langchain
streamlit